Removed warning

+++ src/SFileReadFile.cpp
@@ -197,7 +197,7 @@ static DWORD ReadMpqSectors(TMPQFile * hf, LPBYTE pbBuffer, DWORD dwByteOffset, 
 
                     // Special case (MPQ_2024_v1_300TK2.09p.w3x, file File00010254.blp):
                     // Extracted less than required. Fill the rest with zeros
-                    if(cbOutSector < dwBytesInThisSector)
+                    if((DWORD)(cbOutSector) < dwBytesInThisSector)
                     {
                         memset(pbOutSector + cbOutSector, 0, dwBytesInThisSector - cbOutSector);
                     }
